Feat: Refactor so it is easier when we add new routes w/Router
